pyproject.toml
[project]
name = "futureboosting"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.12.0",
    "arch>=8.0.0",
    "chinese-calendar>=1.11.0",
    "chronos-forecasting>=2.2.2",
    "datasets>=2.17",
    "gluonts>=0.14",
    "holidays>=0.88",
    "lightgbm>=4.6.0",
    "matplotlib>=3.8.0",
    "mlflow>=3.6.0",
    "numpy>=1.26",
    "pandas>=2.1",
    "plotly>=6.5.0",
    "pyarrow>=22.0.0",
    "scikit-learn>=1.0.2",
    "scipy>=1.5",
    "statsmodels>=0.14.5",
    "shap>=0.46.0",
    "swanlab>=0.7.2",
    "timesfm>=1.3.0",
    "tirex-ts>=1.3.0",
    "torch==2.4.0",
    "torchmetrics>=1.8.2",
    "tqdm>=4.67.1",
    "transformers==4.57",
    "uni2ts>=2.0.0",
]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true
